Further basic fixes and clean-up
Removed idea, nbproject folders
Removed aircompressor-0.7.jar; replaced w/gradle dependency
Added slf4j-simple logger
Fixed gradlew line sep
Removed WinRegistry.jar
Further Java clean-up throughout

+++ src/main/java/skyproc/NIF.java
@@ -81,7 +81,7 @@ public class NIF {
         }
         if (!in.getString(20).equals("Gamebryo File Format")) {
             byte first = in.extract(1)[0];
-            if (!in.extractString((int) first, 20).equals("Gamebryo File Format")) {
+            if (!in.extractString(first, 20).equals("Gamebryo File Format")) {
                 throw new BadParameter(fileName + " was not a NIF file.");
             }
         }
